修改draggable名为moveable
