+++ src/app/demo/dragdrop/drap-to-replace/app.component.html
@@ -1,14 +1,14 @@
-<div class="drag-wrap" jigsaw-draggable-h5 jigsaw-droppable-h5>
+<div class="drag-wrap" jigsaw-draggable jigsaw-droppable>
     <div class="com-box box1">
         <button>111</button>
     </div>
 </div>
-<div class="drag-wrap" jigsaw-draggable-h5 jigsaw-droppable-h5>
+<div class="drag-wrap" jigsaw-draggable jigsaw-droppable>
     <div class="com-box box2">
         <button>222</button>
     </div>
 </div>
-<div class="drag-wrap" jigsaw-draggable-h5 jigsaw-droppable-h5>
+<div class="drag-wrap" jigsaw-draggable jigsaw-droppable>
     <div class="com-box box3">
         <button>333</button>
     </div>
