+++ src/app/demo/dragdrop/drap-to-replace/app.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from "@angular/core";
 import {DragToReplaceDemoComponent} from "./app.component";
-import {JigsawH5DraggableModule, JigsawH5DroppableModule} from "jigsaw/component/dragdrop/index";
+import {JigsawDraggableModule, JigsawDroppableModule} from "jigsaw/component/dragdrop/index";
 
 @NgModule({
-    imports: [JigsawH5DraggableModule, JigsawH5DroppableModule],
+    imports: [JigsawDraggableModule, JigsawDroppableModule],
     declarations: [DragToReplaceDemoComponent],
     bootstrap: [DragToReplaceDemoComponent]
 })
