+++ src/jigsaw/component/alert/alert.html
@@ -1,5 +1,5 @@
 <div class="jigsaw-alert" [ngClass]="_$alertClass">
-    <div class="jigsaw-alert-head" jigsaw-draggable draggableAffected=".jigsaw-common-alert">
+    <div class="jigsaw-alert-head" jigsaw-moveable moveableAffected=".jigsaw-common-alert">
         <i class="jigsaw-alert-icon fa "></i>
         <span class="jigsaw-alert-title">{{title | translate}}</span>
         <span class="jigsaw-alert-close fa fa-times" (click)='dispose()'></span>
