+++ src/jigsaw/component/dialog/dialog.html
@@ -1,5 +1,5 @@
 <div class="jigsaw-dialog">
-    <div class="jigsaw-dialog-head" jigsaw-draggable draggableAffected="jigsaw-dialog">
+    <div class="jigsaw-dialog-head" jigsaw-moveable moveableAffected="jigsaw-dialog">
         <ng-content select="[jigsaw-title]"></ng-content>
         <span>{{title}}</span>
         <span class="fa fa-times jigsaw-dialog-close" (click)='dispose()'></span>
