+++ src/jigsaw/component/dragdrop/index.ts
@@ -1,7 +1,14 @@
 import {NgModule} from "@angular/core";
+import {JigsawMoveable} from "./moveable";
 import {JigsawDraggable} from "./draggable";
-import {JigsawH5Draggable} from "./draggable-h5";
-import {JigsawH5Droppable} from "./droppable-h5";
+import {JigsawDroppable} from "./droppable";
+
+@NgModule({
+    declarations: [JigsawMoveable],
+    exports: [JigsawMoveable]
+})
+export class JigsawMoveableModule {
+}
 
 @NgModule({
     declarations: [JigsawDraggable],
@@ -11,20 +18,13 @@ export class JigsawDraggableModule {
 }
 
 @NgModule({
-    declarations: [JigsawH5Draggable],
-    exports: [JigsawH5Draggable]
+    declarations: [JigsawDroppable],
+    exports: [JigsawDroppable]
 })
-export class JigsawH5DraggableModule {
+export class JigsawDroppableModule {
 }
 
-@NgModule({
-    declarations: [JigsawH5Droppable],
-    exports: [JigsawH5Droppable]
-})
-export class JigsawH5DroppableModule {
-}
-
+export * from './moveable';
 export * from './draggable';
-export * from './draggable-h5';
-export * from './droppable-h5';
+export * from './droppable';
