+++ src/jigsaw/module.ts
@@ -7,7 +7,7 @@ import {JigsawCheckBoxModule} from './component/checkbox/index';
 import {JigsawCollapseModule} from './component/collapse/collapse';
 import {JigsawComboSelectModule} from './component/combo-select/index';
 import {JigsawDialogModule} from './component/dialog/dialog';
-import {JigsawDraggableModule, JigsawH5DraggableModule, JigsawH5DroppableModule} from './component/dragdrop/index';
+import {JigsawMoveableModule, JigsawDraggableModule, JigsawDroppableModule} from './component/dragdrop/index';
 import {JigsawGraphModule} from './component/graph/index';
 import {JigsawInputModule} from './component/input/input';
 import {JigsawLoadingModule} from './component/loading/loading';
@@ -34,9 +34,9 @@ const JIGSAW_MODULE = [
     JigsawCollapseModule,
     JigsawComboSelectModule,
     JigsawDialogModule,
+    JigsawMoveableModule,
     JigsawDraggableModule,
-    JigsawH5DraggableModule,
-    JigsawH5DroppableModule,
+    JigsawDroppableModule,
     JigsawGraphModule,
     JigsawInputModule,
     JigsawLoadingModule,
